Add set_budget and check_budget functions

+++ src/main/java/seedu/budget/BillsBudget.java
@@ -3,7 +3,7 @@ package seedu.budget;
 import seedu.entry.ExpenseCategory;
 
 public class BillsBudget extends Budget {
-    public BillsBudget(int limit) {
+    public BillsBudget(double limit) {
         this.category = ExpenseCategory.BILLS;
         this.name = ExpenseCategory.BILLS.toString();
         this.limit = limit;
