+++ src/main/java/seedu/budget/EntertainmentBudget.java
@@ -3,7 +3,7 @@ package seedu.budget;
 import seedu.entry.ExpenseCategory;
 
 public class EntertainmentBudget extends Budget {
-    public EntertainmentBudget(int limit) {
+    public EntertainmentBudget(double limit) {
         this.category = ExpenseCategory.ENTERTAINMENT;
         this.name = ExpenseCategory.ENTERTAINMENT.toString();
         this.limit = limit;
