+++ src/main/java/seedu/budget/FoodBudget.java
@@ -4,7 +4,7 @@ import seedu.entry.ExpenseCategory;
 
 public class FoodBudget extends Budget {
 
-    public FoodBudget(int limit) {
+    public FoodBudget(double limit) {
         this.category = ExpenseCategory.FOOD;
         this.name = ExpenseCategory.FOOD.toString();
         this.limit = limit;
