+++ src/main/java/seedu/budget/MedicalBudget.java
@@ -4,7 +4,7 @@ import seedu.entry.ExpenseCategory;
 
 public class MedicalBudget extends Budget {
 
-    public MedicalBudget(int limit) {
+    public MedicalBudget(double limit) {
         this.category = ExpenseCategory.MEDICAL;
         this.name = ExpenseCategory.MEDICAL.toString();
         this.limit = limit;
