+++ src/main/java/seedu/budget/MiscBudget.java
@@ -3,7 +3,7 @@ package seedu.budget;
 import seedu.entry.ExpenseCategory;
 
 public class MiscBudget extends Budget {
-    public MiscBudget(int limit) {
+    public MiscBudget(double limit) {
         this.category = ExpenseCategory.MISC;
         this.name = ExpenseCategory.MISC.toString();
         this.limit = limit;
