+++ src/main/java/seedu/budget/TransportBudget.java
@@ -4,7 +4,7 @@ import seedu.entry.ExpenseCategory;
 
 public class TransportBudget extends Budget {
 
-    public TransportBudget(int limit) {
+    public TransportBudget(double limit) {
         this.category = ExpenseCategory.TRANSPORT;
         this.name = ExpenseCategory.TRANSPORT.toString();
         this.limit = limit;
